compile fix for system that do not have size_t defined as unsigned long
git-svn-id: https://pcapsplit.svn.sourceforge.net/svnroot/pcapsplit@56 9b7ec49d-eebb-4683-924c-5cfa09519806

+++ size_dumper.c
@@ -42,7 +42,7 @@ struct size_dumper_data {
 void createNewFile(struct size_dumper_data* data)
 {
 	snprintf(data->dump_filename, MAX_FILENAME, "%s.%lu",
-		data->base_filename, data->number);
+		data->base_filename, (unsigned long)data->number);
 	//TODO: use DLT from source file
 	data->out = pcap_open_dead(DLT_EN10MB, 65535);
 	data->dumper = pcap_dump_open(data->out, data->dump_filename);
